requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
greenlet==3.3.1
idna==3.11
numpy==2.4.2
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
pyyaml==6.0.2

